requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
xgboost
statsmodels
imbalanced-learn
ipykernel
python-dotenv
fastapi
structlog
tavily-python
pydantic
wikipedia
uvicorn
httpx
aiofiles
sqlalchemy
bcrypt
reportlab
python-docx
passlib
shap
pyyaml

-e .